Move more stuff to bytecode.rs
Now it's more clear that all data structures such as Target is only
relevant to compiler.
