+++ parser/src/compiler.rs
@@ -1,289 +1,14 @@
-use std::{
-    cell::RefCell,
-    collections::HashMap,
-    io::{Read, Write},
-    rc::Rc,
-};
+use std::{cell::RefCell, collections::HashMap, io::Write, rc::Rc};
 
 use crate::{
-    bytecode::{Instruction, OpCode},
-    interpreter::{
-        eval, s_hex_string, s_len, s_print, s_push, s_type, EvalContext, EvalError, RunResult,
+    bytecode::{
+        std_functions, Bytecode, BytecodeArg, FnBytecode, FnProto, Instruction, NativeFn, OpCode,
     },
-    parser::{ExprEnum, Expression, ReadError, Statement},
+    interpreter::{eval, EvalContext, RunResult},
+    parser::{ExprEnum, Expression, Statement},
     value::ArrayInt,
     TypeDecl, Value,
 };
-
-macro_rules! dbg_println {
-    ($($rest:tt)*) => {{
-        #[cfg(debug_assertions)]
-        std::println!($($rest)*)
-    }}
-}
-
-fn write_str(s: &str, writer: &mut impl Write) -> std::io::Result<()> {
-    writer.write_all(&s.len().to_le_bytes())?;
-    writer.write_all(&s.as_bytes())?;
-    Ok(())
-}
-
-fn read_str(reader: &mut impl Read) -> Result<String, ReadError> {
-    let mut len = [0u8; std::mem::size_of::<usize>()];
-    reader.read_exact(&mut len)?;
-    let len = usize::from_le_bytes(len);
-    let mut buf = vec![0u8; len];
-    reader.read_exact(&mut buf)?;
-    Ok(String::from_utf8(buf)?)
-}
-
-fn write_bool(b: bool, writer: &mut impl Write) -> std::io::Result<()> {
-    writer.write_all(&[if b { 1u8 } else { 0u8 }])
-}
-
-fn read_bool(reader: &mut impl Read) -> Result<bool, ReadError> {
-    let mut buf = [0u8; 1];
-    reader.read_exact(&mut buf)?;
-    Ok(buf[0] != 0)
-}
-
-fn write_opt_value(value: &Option<Value>, writer: &mut impl Write) -> std::io::Result<()> {
-    write_bool(value.is_some(), writer)?;
-    if let Some(value) = value {
-        value.serialize(writer)?;
-    }
-    Ok(())
-}
-
-fn read_opt_value(reader: &mut impl Read) -> Result<Option<Value>, ReadError> {
-    let has_value = read_bool(reader)?;
-    Ok(if has_value {
-        Some(Value::deserialize(reader)?)
-    } else {
-        None
-    })
-}
-
-pub type NativeFn = Box<dyn Fn(&[Value]) -> Result<Value, EvalError>>;
-
-pub(crate) enum FnProto {
-    Code(FnBytecode),
-    Native(NativeFn),
-}
-
-impl FnProto {
-    pub fn args(&self) -> &[BytecodeArg] {
-        match self {
-            Self::Code(bytecode) => &bytecode.args,
-            _ => &[],
-        }
-    }
-}
-
-pub struct Bytecode {
-    pub(crate) functions: HashMap<String, FnProto>,
-}
-
-impl Bytecode {
-    /// Add a user-application provided native function to this bytecode.
-    pub fn add_ext_fn(
-        &mut self,
-        name: String,
-        f: Box<dyn Fn(&[Value]) -> Result<Value, EvalError>>,
-    ) {
-        self.functions.insert(name, FnProto::Native(f));
-    }
-
-    pub fn add_std_fn(&mut self) {
-        std_functions(&mut |name, f| self.add_ext_fn(name, f));
-    }
-
-    pub fn write(&self, writer: &mut impl Write) -> std::io::Result<()> {
-        writer.write_all(
-            &self
-                .functions
-                .iter()
-                .filter(|f| matches!(f.1, FnProto::Code(_)))
-                .count()
-                .to_le_bytes(),
-        )?;
-        for (fname, func) in self.functions.iter() {
-            if let FnProto::Code(func) = func {
-                write_str(fname, writer)?;
-                func.write(writer)?;
-            }
-        }
-        Ok(())
-    }
-
-    pub fn read(reader: &mut impl Read) -> Result<Self, ReadError> {
-        let mut len = [0u8; std::mem::size_of::<usize>()];
-        reader.read_exact(&mut len)?;
-        let len = usize::from_le_bytes(len);
-        let ret = Bytecode {
-            functions: (0..len)
-                .map(|_| -> Result<(String, FnProto), ReadError> {
-                    Ok((read_str(reader)?, FnProto::Code(FnBytecode::read(reader)?)))
-                })
-                .collect::<Result<HashMap<_, _>, ReadError>>()?,
-        };
-        dbg_println!("loaded {} functions", ret.functions.len());
-        let loaded_fn = ret
-            .functions
-            .iter()
-            .find(|(name, _)| *name == "")
-            .ok_or(ReadError::NoMainFound)?;
-        if let FnProto::Code(ref _code) = loaded_fn.1 {
-            dbg_println!("instructions: {:#?}", _code.instructions);
-        }
-        Ok(ret)
-    }
-}
-
-/// Add standard common functions, such as `print`, `len` and `push`, to this bytecode.
-pub fn std_functions(
-    f: &mut impl FnMut(String, Box<dyn Fn(&[Value]) -> Result<Value, EvalError>>),
-) {
-    f("print".to_string(), Box::new(s_print));
-    f(
-        "puts".to_string(),
-        Box::new(|values: &[Value]| -> Result<Value, EvalError> {
-            print!(
-                "{}",
-                values.iter().fold("".to_string(), |acc, cur: &Value| {
-                    if acc.is_empty() {
-                        cur.to_string()
-                    } else {
-                        acc + &cur.to_string()
-                    }
-                })
-            );
-            Ok(Value::I64(0))
-        }),
-    );
-    f("type".to_string(), Box::new(&s_type));
-    f("len".to_string(), Box::new(s_len));
-    f("push".to_string(), Box::new(s_push));
-    f("hex_string".to_string(), Box::new(s_hex_string));
-}
-
-#[derive(Debug, Clone)]
-pub struct BytecodeArg {
-    name: String,
-    init: Option<Value>,
-}
-
-#[derive(Debug, Clone)]
-pub struct FnBytecode {
-    pub(crate) literals: Vec<Value>,
-    pub(crate) args: Vec<BytecodeArg>,
-    pub(crate) instructions: Vec<Instruction>,
-    pub(crate) stack_size: usize,
-}
-
-impl FnBytecode {
-    /// Create a placeholder entry that will be filled later.
-    fn proto(args: Vec<String>) -> Self {
-        Self {
-            literals: vec![],
-            args: args
-                .into_iter()
-                .map(|arg| BytecodeArg {
-                    name: arg,
-                    init: None,
-                })
-                .collect(),
-            instructions: vec![],
-            stack_size: 0,
-        }
-    }
-
-    fn push_inst(&mut self, op: OpCode, arg0: u8, arg1: u16) -> usize {
-        let ret = self.instructions.len();
-        self.instructions.push(Instruction::new(op, arg0, arg1));
-        ret
-    }
-
-    pub fn write(&self, writer: &mut impl Write) -> std::io::Result<()> {
-        writer.write_all(&self.stack_size.to_le_bytes())?;
-        writer.write_all(&self.literals.len().to_le_bytes())?;
-        for literal in &self.literals {
-            literal.serialize(writer)?;
-        }
-        writer.write_all(&self.args.len().to_le_bytes())?;
-        for arg in &self.args {
-            write_str(&arg.name, writer)?;
-            write_opt_value(&arg.init, writer)?;
-        }
-        writer.write_all(&self.instructions.len().to_le_bytes())?;
-        for inst in &self.instructions {
-            inst.serialize(writer)?;
-        }
-        Ok(())
-    }
-
-    pub fn read(reader: &mut impl Read) -> Result<Self, ReadError> {
-        let mut stack_size = [0u8; std::mem::size_of::<usize>()];
-        reader.read_exact(&mut stack_size)?;
-
-        let mut literals = [0u8; std::mem::size_of::<usize>()];
-        reader.read_exact(&mut literals)?;
-        let literals = usize::from_le_bytes(literals);
-        let literals = (0..literals)
-            .map(|_| Value::deserialize(reader))
-            .collect::<Result<Vec<_>, _>>()?;
-
-        let mut args = [0u8; std::mem::size_of::<usize>()];
-        reader.read_exact(&mut args)?;
-        let num_args = usize::from_le_bytes(args);
-        let args = (0..num_args)
-            .map(|_| -> Result<_, ReadError> {
-                let name = read_str(reader)?;
-                let init = read_opt_value(reader)?;
-                Ok(BytecodeArg { name, init })
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-
-        let mut instructions = [0u8; std::mem::size_of::<usize>()];
-        reader.read_exact(&mut instructions)?;
-        let instructions = usize::from_le_bytes(instructions);
-        let instructions = (0..instructions)
-            .map(|_| Instruction::deserialize(reader))
-            .collect::<Result<Vec<_>, _>>()?;
-        Ok(Self {
-            literals,
-            args,
-            instructions,
-            stack_size: usize::from_le_bytes(stack_size),
-        })
-    }
-
-    pub fn disasm(&self, f: &mut impl std::io::Write) -> Result<(), std::io::Error> {
-        writeln!(f, "Stack size: {}", self.stack_size)?;
-        writeln!(f, "Literals({}):", self.literals.len())?;
-        for (i, literal) in self.literals.iter().enumerate() {
-            writeln!(f, "  [{}] {}", i, literal)?;
-        }
-        writeln!(f, "Args({}):", self.args.len())?;
-        for (i, arg) in self.args.iter().enumerate() {
-            writeln!(f, "  [{}] {} = {:?}", i, arg.name, arg.init)?;
-        }
-        writeln!(f, "Instructions({}):", self.instructions.len())?;
-        for (i, inst) in self.instructions.iter().enumerate() {
-            match inst.op {
-                OpCode::LoadLiteral => {
-                    if let Some(literal) = self.literals.get(inst.arg0 as usize) {
-                        writeln!(f, "  [{}] {} ({:?})", i, inst, literal)?;
-                    } else {
-                        writeln!(f, "  [{}] {} ? (Literal index out of bound)", i, inst)?;
-                    }
-                }
-                _ => writeln!(f, "  [{}] {}", i, inst)?,
-            }
-        }
-        Ok(())
-    }
-}
 
 #[derive(Debug, Clone, Copy)]
 enum Target {
@@ -571,10 +296,7 @@ fn emit_stmts(stmts: &[Statement], compiler: &mut Compiler) -> Result<Option<usi
                         } else {
                             None
                         };
-                        Ok(BytecodeArg {
-                            name: arg.name.to_owned(),
-                            init,
-                        })
+                        Ok(BytecodeArg::new(arg.name.to_owned(), init))
                     })
                     .collect::<Result<_, _>>()?;
                 dbg_println!("FnDecl actual args: {:?} fn_args: {:?}", a_args, fn_args);
