+++ parser/src/vm.rs
@@ -3,12 +3,13 @@
 use std::collections::HashMap;
 
 use crate::{
+    bytecode::{Bytecode, FnBytecode, FnProto, OpCode},
     interpreter::{
         binary_op, binary_op_int, binary_op_str, coerce_f64, coerce_i64, coerce_type, truthy,
         EvalError,
     },
     type_decl::TypeDecl,
-    Bytecode, FnBytecode, FnProto, OpCode, Value,
+    Value,
 };
 
 macro_rules! dbg_println {
